WiFiSSLClient: fix Wsign-compare

+++ libraries/WiFiS3/src/WiFiSSLClient.cpp
@@ -148,7 +148,7 @@ int WiFiSSLClient::_read() {
       modem.read_using_size();
       if(modem.write(string(PROMPT(_SSLCLIENTRECEIVE)),res, "%s%d,%d\r\n" , CMD_WRITE(_SSLCLIENTRECEIVE), _sock, size)) {
          if(res.size() > 0) {
-            for(int i = 0, rv = 0; i < size && i < res.size(); i++) {
+            for(uint32_t i = 0, rv = 0; i < size && i < res.size(); i++) {
                rx_buffer->store((uint8_t)res[i]);
                rv++;
             }
@@ -185,7 +185,7 @@ int WiFiSSLClient::read(uint8_t *buf, size_t size) {
    read_if_needed(size);
    int rv = 0;
    bool go_on = true;
-   for(int i = 0; i < size && go_on; i++) {
+   for(size_t i = 0; i < size && go_on; i++) {
       bool is_read = false;
       *(buf+i) = rx_buffer->read(&is_read);
       if(is_read) {
